fix: strategic AI crashing
The when there were boxes with 3 lines surrounding them the AI tried to look at key 3 in its lookup hashmap...

Which did not exist since I was thinking in terms of zero-indexing when I first did wrote the sizing of that lookup.

What's strange though is I would have thought that it would segmentation fault or some other.

But since the hashmap was undersized it kept looping through all the keys to find an empty slot that would never come.